refactor: make settings the messaging configuration home
