+++ docs/pages/platform-ms-teams.md
@@ -3,7 +3,7 @@ title: MS Teams
 category: Agents
 order: 8
 description: Connect Archestra agents to Microsoft Teams channels
-lastUpdated: 2026-08-27
+lastUpdated: 2026-08-28
 ---
 
 <!-- Renaming/deleting this file? Add a redirect in docs/redirects.json. -->
