+++ platform/frontend/src/app/settings/messaging-channels/_components/channel-instructions-dialog.tsx
