+++ platform/frontend/src/app/settings/messaging-channels/_components/channels-section.tsx
@@ -10,7 +10,6 @@ import {
   Inbox,
   MessageSquareText,
   Plus,
-  Search,
   TriangleAlert,
   X,
 } from "lucide-react";
