+++ platform/frontend/src/app/settings/messaging-channels/_components/credential-field.tsx
