+++ platform/frontend/src/app/settings/messaging-channels/_components/external-docs-link.test.tsx
