+++ platform/frontend/src/app/settings/messaging-channels/_components/external-docs-link.tsx
