+++ platform/frontend/src/app/settings/messaging-channels/_components/setup-step.tsx
