+++ platform/frontend/src/app/settings/messaging-channels/_components/types.ts
