+++ platform/frontend/src/app/settings/messaging-channels/_components/use-trigger-statuses.ts
@@ -1,4 +1,3 @@
-import { useInternalAgents } from "@/lib/agent.query";
 import { useChatOpsStatus } from "@/lib/chatops/chatops.query";
 import { useIncomingEmailStatus } from "@/lib/chatops/incoming-email.query";
 import config from "@/lib/config/config";
@@ -15,10 +14,6 @@ export function useTriggerStatuses() {
     useIncomingEmailStatus();
   const { data: chatApiKeys = [], isLoading: apiKeysLoading } =
     useLlmProviderApiKeys();
-  const { data: internalAgents, isLoading: agentsLoading } = useInternalAgents({
-    enabled: true,
-  });
-
   const channelCatalog = useMessagingChannelCatalog();
 
   const hasLlmKey = chatApiKeys.length > 0;
@@ -54,33 +49,23 @@ export function useTriggerStatuses() {
   const emailActive =
     !!configData?.features.incomingEmail?.enabled && !!emailStatus?.isActive;
 
-  // A2A surfaces existing agents over the A2A protocol — no provider config
-  // step, so it's "active" whenever there's at least one agent to expose.
-  const a2aActive = (internalAgents?.length ?? 0) > 0;
-
-  // Landing candidates are the channels that actually have a tab: one the
-  // admins turned off, or that this deployment never enabled, would otherwise
-  // be a redirect straight onto a dead page.
-  const triggers = (
+  // Landing candidates are the providers that actually have a tab. A provider
+  // that admins turned off, or that this deployment never enabled, must not be
+  // selected as the Settings landing page.
+  const providerTriggers = (
     [
       { id: "ms-teams", active: msTeamsActive, available: true },
       { id: "slack", active: slackActive, available: true },
       { id: "telegram", active: telegramActive, available: telegramAvailable },
       { id: "email", active: emailActive, available: true },
-      { id: "a2a", active: a2aActive, available: true },
     ] as const
   )
     .filter(({ id, available }) => available && !channelCatalog.isHidden(id))
     .map(({ id, active }) => ({
       id,
       active,
-      href: `/messaging-channels/${id}`,
+      href: `/settings/messaging-channels/${id}`,
     }));
-  // Null when every channel is off — there is nowhere to send anyone, so the
-  // index route stays put and renders the empty state instead.
-  const firstActiveHref =
-    triggers.find((t) => t.active)?.href ?? triggers[0]?.href ?? null;
-  const providerTriggers = triggers.filter(({ id }) => id !== "a2a");
   const firstProviderHref =
     providerTriggers.find((trigger) => trigger.active)?.href ??
     providerTriggers[0]?.href ??
@@ -92,14 +77,8 @@ export function useTriggerStatuses() {
     telegram: telegramActive,
     telegramAvailable,
     email: emailActive,
-    a2a: a2aActive,
-    firstActiveHref,
     firstProviderHref,
     isLoading:
-      chatOpsLoading ||
-      featuresLoading ||
-      emailLoading ||
-      apiKeysLoading ||
-      agentsLoading,
+      chatOpsLoading || featuresLoading || emailLoading || apiKeysLoading,
   };
 }
