+++ platform/frontend/src/app/settings/messaging-channels/email/agent-email-settings-dialog.tsx
