+++ platform/frontend/src/app/settings/messaging-channels/email/email-trigger.utils.ts
