+++ platform/frontend/src/app/settings/messaging-channels/layout.tsx
@@ -6,10 +6,10 @@ import {
 } from "@archestra/shared";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { useTriggerStatuses } from "@/app/messaging-channels/_components/use-trigger-statuses";
 import { ChannelIcon } from "@/components/channel-icon";
 import { useMessagingChannelCatalog } from "@/lib/integration-overrides";
 import { cn } from "@/lib/utils";
+import { useTriggerStatuses } from "./_components/use-trigger-statuses";
 
 const CHAT_PROVIDER_IDS = [
   "ms-teams",
